+++ packages/headless/src/modules/user/rest/dto/changePassword.dto.ts
